backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
websockets==12.0
httpx==0.26.0
langgraph==0.0.20
openai==1.10.0
tenacity==8.2.3
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
redis==5.0.1
python-multipart==0.0.6
